requirements.txt
Flask==3.0.0
reportlab==4.0.7
Pillow==10.1.0
openpyxl==3.1.2
mysql-connector-python==8.2.0
